refactor(draft-packages): rename scss files to use .module.scss
- avatar
- badge
- card
- collapsible
- divider
- empty-state
- filter-menu-button
