+++ draft-packages/badge/KaizenDraft/Badge/Badge.module.scss
@@ -2,7 +2,7 @@
 @import "~@kaizen/design-tokens/sass/typography";
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/spacing";
-@import "./decision-tokens.scss";
+@import "./variables";
 
 $small: $spacing-md;
