+++ draft-packages/badge/KaizenDraft/Badge/_variables.scss
