--- draft-packages/badge/docs/styles.module.scss
@@ -1,30 +1,0 @@
-@import "~@kaizen/component-library/styles/responsive";
-
-.badgeSection {
-  display: flex;
-  justify-content: flex-start;
-  flex-direction: row;
-  align-items: flex-start;
-  margin-bottom: 3rem;
-
-  @include ca-media-mobile {
-    flex-direction: column;
-    margin-bottom: 1rem;
-  }
-}
-
-.badgeSize {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin-right: 5rem;
-
-  h3 {
-    margin-bottom: 0.5rem;
-  }
-
-  @include ca-media-mobile {
-    flex-direction: column;
-    margin-top: 1rem;
-  }
-}
