+++ draft-packages/collapsible/KaizenDraft/Collapsible/Collapsible.module.scss
