+++ draft-packages/collapsible/KaizenDraft/Collapsible/Collapsible.tsx
@@ -7,7 +7,7 @@ import { Heading } from "@kaizen/typography"
 import chevronUp from "@kaizen/component-library/icons/chevron-up.icon.svg"
 import chevronDown from "@kaizen/component-library/icons/chevron-down.icon.svg"
 import { Sticky } from "./CollapsibleGroup"
-import styles from "./styles.scss"
+import styles from "./Collapsible.module.scss"
 
 type Variant = "default" | "clear"
