+++ draft-packages/collapsible/KaizenDraft/Collapsible/CollapsibleGroup.module.scss
@@ -1,0 +1,10 @@
+@import "~@kaizen/design-tokens/sass/border";
+@import "~@kaizen/design-tokens/sass/shadow";
+
+// We need a full border radius on this container element, then have classes
+// beneath to toggle nested borders off and on for different use cases.
+.container {
+  background-color: white;
+  box-shadow: $shadow-small-box-shadow;
+  border-radius: $border-borderless-border-radius;
+}
