+++ draft-packages/collapsible/docs/Collapsible.stories.module.scss
