+++ draft-packages/collapsible/docs/Collapsible.stories.tsx
@@ -4,7 +4,7 @@ import { Paragraph } from "@kaizen/typography"
 import { Collapsible, CollapsibleGroup } from "@kaizen/draft-collapsible"
 import translationIcon from "@kaizen/component-library/icons/translation.icon.svg"
 import { CATEGORIES } from "../../../storybook/constants"
-import styles from "./Collapsible.stories.scss"
+import styles from "./Collapsible.stories.module.scss"
 
 const ListItem = ({ children }: { children: JSX.Element }) => (
   <div className={styles.listItem}>{children}</div>
