+++ draft-packages/empty-state/KaizenDraft/EmptyState/EmptyState.module.scss
@@ -1,7 +1,7 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 @import "~@kaizen/design-tokens/sass/color";
 @import "~@kaizen/design-tokens/sass/border";
-@import "./responsive";
+@import "./mixins";
 
 .container {
   display: flex;
