+++ draft-packages/empty-state/KaizenDraft/EmptyState/_mixins.scss
